+++ src/model/ConvoDraft.ts
@@ -11,16 +11,12 @@ export type ConvoDraft = {
 
 export type UploadingAttach = GenericUploadingAttach<'Photo'>
 
-type GenericUploadingAttach<Kind extends string> = {
+type GenericUploadingAttach<Kind extends Attach.SingleAttach['kind']> = {
   kind: Kind
   file: File
   progress: number
   failed: boolean
 }
-
-export type AttachPreview =
-  | { kind: 'Attach', attach: Attach.SingleAttach }
-  | { kind: 'UploadingAttach', attach: UploadingAttach }
 
 const emptyDraft = (): ConvoDraft => ({
   text: '',
@@ -38,12 +34,8 @@ export function get(peerId: Peer.Id): ConvoDraft {
   return getMapValueWithDefaults(drafts, peerId, emptyDraft())
 }
 
-export function isEmpty(draft: ConvoDraft, countUploading = true) {
-  return (
-    draft.text === '' &&
-    Attach.kindsCount(draft.attaches) === 0 &&
-    (!countUploading || draft.uploadingAttaches.length === 0)
-  )
+export function isEmpty(draft: ConvoDraft) {
+  return draft.text.trim() === '' && Attach.kindsCount(draft.attaches) === 0
 }
 
 export function reset(draft: ConvoDraft) {
@@ -64,29 +56,4 @@ export function removeUploadingAttach(draft: ConvoDraft, uploadingAttach: Upload
   if (index !== -1) {
     draft.uploadingAttaches.splice(index, 1)
   }
-}
-
-export function getAttachesList(draft: ConvoDraft): AttachPreview[] {
-  const attaches: AttachPreview[] = []
-  const draftAttachesValues = Object.values(draft.attaches)
-
-  for (const attach of draftAttachesValues) {
-    const attachesArray = Array.isArray(attach) ? attach : [attach]
-    const uploadingAttaches = draft.uploadingAttaches.filter((uploadingAttach) => (
-      uploadingAttach.kind === attachesArray[0].kind
-    ))
-
-    attaches.push(
-      ...attachesArray.map((attach) => ({ kind: 'Attach' as const, attach })),
-      ...uploadingAttaches.map((attach) => ({ kind: 'UploadingAttach' as const, attach }))
-    )
-  }
-
-  const restUploadingAttaches = draft.uploadingAttaches
-    .filter((uploadingAttach) => !attaches.some(({ attach }) => attach === uploadingAttach))
-    .map((attach) => ({ kind: 'UploadingAttach' as const, attach }))
-
-  attaches.push(...restUploadingAttaches)
-
-  return attaches
 }
